feat: remove duplicates in place
Learning the writer seeker technique and I found out
I could check for uniques by checking for the
previous value, not the next value, because of a
loop invariant. After a writer moves the values
left behind are unique. Therefore, if the next value
is the same it must not be unique and the seeker
searches for the next unique value while the writer
holds the place to write the unique value.